+++ Tests/MapboxSearchTests/Use Cases/Details/SearchEngineTests+Details.swift
@@ -83,7 +83,7 @@ extension SearchEngineTests {
         let searchOptions = SearchOptions(
             limit: 100,
             origin: CLLocationCoordinate2D(latitude: 38.902309, longitude: -77.029129),
-            filterTypes: [.poi]
+            filterQueryTypes: [.poi]
         )
 
         searchEngine.search(query: "dolcezza gelato", options: searchOptions)
@@ -160,7 +160,7 @@ extension SearchEngineTests {
         let searchOptions = SearchOptions(
             limit: 100,
             origin: CLLocationCoordinate2D(latitude: 38.902309, longitude: -77.029129),
-            filterTypes: [.poi]
+            filterQueryTypes: [.poi]
         )
 
         searchEngine.search(query: "planet word", options: searchOptions)
@@ -218,7 +218,7 @@ extension SearchEngineTests {
         let searchOptions = SearchOptions(
             limit: 100,
             origin: CLLocationCoordinate2D(latitude: 38.902309, longitude: -77.029129),
-            filterTypes: [.poi]
+            filterQueryTypes: [.poi]
         )
 
         searchEngine.search(query: "planet word", options: searchOptions)